requirements.txt
numpy==1.16.4
opencv-contrib-python==4.1.0.25
tensorflow-gpu==1.13.1
